refactor(contracts): split tx-interfaces into call (kept) and governance/tx-interfaces
