+++ packages/contracts/src/test/deploy-contract.test.ts
@@ -36,7 +36,10 @@ vi.mock('../submit-deploy-tx', () => ({
 }));
 
 vi.mock('../tx-interfaces', () => ({
-  createCircuitCallTxInterface: vi.fn().mockReturnValue({ call: 'mock-call-interface' }),
+  createCircuitCallTxInterface: vi.fn().mockReturnValue({ call: 'mock-call-interface' })
+}));
+
+vi.mock('../governance/tx-interfaces', () => ({
   createCircuitMaintenanceTxInterfaces: vi.fn().mockReturnValue({ maintenance: 'mock-maintenance-interfaces' }),
   createContractMaintenanceTxInterface: vi.fn().mockReturnValue({ contractMaintenance: 'mock-contract-maintenance' })
 }));
