+++ packages/contracts/src/test/tx-interfaces.test.ts
@@ -13,15 +13,11 @@
  * limitations under the License.
  */
 
-import type { VerifierKey } from '@midnight-ntwrk/midnight-js-types';
 import { beforeEach, describe, expect, it, vi } from 'vitest';
 
 import {
   createCallTxOptions,
   createCircuitCallTxInterface,
-  createCircuitMaintenanceTxInterface,
-  createCircuitMaintenanceTxInterfaces,
-  createContractMaintenanceTxInterface,
 } from '../tx-interfaces';
 import {
   createMockCompiledContract,
@@ -32,9 +28,6 @@ import {
 } from './test-mocks';
 
 vi.mock('../submit-call-tx');
-vi.mock('../governance/submit-insert-vk-tx');
-vi.mock('../governance/submit-remove-vk-tx');
-vi.mock('../governance/submit-replace-authority-tx');
 
 describe('tx-interfaces', () => {
   let mockCompiledContract: ReturnType<typeof createMockCompiledContract>;
@@ -242,98 +235,6 @@ describe('tx-interfaces', () => {
         expect.not.objectContaining({ args: expect.anything() }),
         txCtx
       );
-    });
-  });
-
-  describe('createCircuitMaintenanceTxInterface', () => {
-    it('should create circuit maintenance interface', () => {
-      const circuitId = 'testCircuit';
-      const maintenanceInterface = createCircuitMaintenanceTxInterface(
-        mockProviders,
-        circuitId,
-        mockCompiledContract,
-        mockContractAddress
-      );
-
-      expect(maintenanceInterface).toHaveProperty('removeVerifierKey');
-      expect(maintenanceInterface).toHaveProperty('insertVerifierKey');
-      expect(typeof maintenanceInterface.removeVerifierKey).toBe('function');
-      expect(typeof maintenanceInterface.insertVerifierKey).toBe('function');
-    });
-
-    it('should call removeVerifierKey function', async () => {
-      const circuitId = 'testCircuit';
-      const maintenanceInterface = createCircuitMaintenanceTxInterface(
-        mockProviders,
-        circuitId,
-        mockCompiledContract,
-        mockContractAddress
-      );
-
-      const { submitRemoveVerifierKeyTx } = await import('../governance/submit-remove-vk-tx');
-      vi.mocked(submitRemoveVerifierKeyTx).mockResolvedValue(mockFinalizedTxData);
-
-      const result = await maintenanceInterface.removeVerifierKey();
-
-      expect(submitRemoveVerifierKeyTx).toHaveBeenCalledWith(
-        mockProviders,
-        mockCompiledContract,
-        mockContractAddress,
-        circuitId
-      );
-      expect(result).toBe(mockFinalizedTxData);
-    });
-
-    it('should call insertVerifierKey function', async () => {
-      const circuitId = 'testCircuit';
-      const mockVerifierKey = new Uint8Array(32) as VerifierKey;
-      const maintenanceInterface = createCircuitMaintenanceTxInterface(
-        mockProviders,
-        circuitId,
-        mockCompiledContract,
-        mockContractAddress
-      );
-
-      const { submitInsertVerifierKeyTx } = await import('../governance/submit-insert-vk-tx');
-      vi.mocked(submitInsertVerifierKeyTx).mockResolvedValue(mockFinalizedTxData);
-
-      const result = await maintenanceInterface.insertVerifierKey(mockVerifierKey);
-
-      expect(submitInsertVerifierKeyTx).toHaveBeenCalledWith(
-        mockProviders,
-        mockCompiledContract,
-        mockContractAddress,
-        circuitId,
-        mockVerifierKey
-      );
-      expect(result).toBe(mockFinalizedTxData);
-    });
-  });
-
-  describe('createCircuitMaintenanceTxInterfaces', () => {
-    it('should create maintenance interfaces for all circuits', () => {
-      const interfaces = createCircuitMaintenanceTxInterfaces(
-        mockProviders,
-        mockCompiledContract,
-        mockContractAddress
-      );
-
-      expect(interfaces).toHaveProperty('testCircuit');
-      expect(interfaces.testCircuit).toHaveProperty('removeVerifierKey');
-      expect(interfaces.testCircuit).toHaveProperty('insertVerifierKey');
-    });
-  });
-
-  describe('createContractMaintenanceTxInterface', () => {
-    it('should create contract maintenance interface', () => {
-      const contractInterface = createContractMaintenanceTxInterface(
-        mockProviders,
-        mockCompiledContract,
-        mockContractAddress
-      );
-
-      expect(contractInterface).toBeDefined();
-      expect(contractInterface).toBeTypeOf('object');
     });
   });
 });
