+++ packages/contracts/src/tx-interfaces.ts
@@ -15,19 +15,12 @@
 
 import { type CompiledContract, ContractExecutable } from '@midnight-ntwrk/midnight-js-protocol/compact-js';
 import type { Contract } from '@midnight-ntwrk/midnight-js-protocol/compact-js/effect/Contract';
-import type { SigningKey } from '@midnight-ntwrk/midnight-js-protocol/compact-runtime';
 import type { CoinPublicKey, ContractAddress, EncPublicKey } from '@midnight-ntwrk/midnight-js-protocol/ledger';
-import {
-  type FinalizedTxData,
-  type PrivateStateId,
-  type VerifierKey} from '@midnight-ntwrk/midnight-js-types';
+import { type PrivateStateId } from '@midnight-ntwrk/midnight-js-types';
 import { assertIsContractAddress } from '@midnight-ntwrk/midnight-js-utils';
 
 import { type CallResult } from './call';
 import { type ContractProviders } from './contract-providers';
-import { submitInsertVerifierKeyTx } from './governance/submit-insert-vk-tx';
-import { submitRemoveVerifierKeyTx } from './governance/submit-remove-vk-tx';
-import { submitReplaceAuthorityTx } from './governance/submit-replace-authority-tx';
 import { submitCallTx } from './submit-call-tx';
 import * as Transaction from './transaction';
 import type { FinalizedCallTxData } from './tx-model';
@@ -103,107 +96,4 @@ export const createCircuitCallTxInterface = <C extends Contract.Any>(
     }),
     {}
   ) as CircuitCallTxInterface<C>;
-};
-
-/**
- * An interface for creating maintenance transactions for a specific circuit defined in a
- * given contract.
- */
-export type CircuitMaintenanceTxInterface = {
-  /**
-   * Constructs and submits a transaction that removes the current verifier key stored
-   * on the blockchain for this circuit at this contract's address.
-   */
-  removeVerifierKey(): Promise<FinalizedTxData>;
-  /**
-   * Constructs and submits a transaction that adds a new verifier key to the
-   * blockchain for this circuit at this contract's address.
-   *
-   * @param newVk The new verifier key to add for this circuit.
-   */
-  insertVerifierKey(newVk: VerifierKey): Promise<FinalizedTxData>;
-}
-
-/**
- * Creates a {@link CircuitMaintenanceTxInterface}.
- *
- * @param providers The providers to use to create and submit transactions.
- * @param circuitId The circuit ID the interface is for.
- * @param contractAddress The address of the deployed contract for which this
- *                        interface is being created.
- */
-export const createCircuitMaintenanceTxInterface = <C extends Contract.Any, PCK extends Contract.ProvableCircuitId<C>>(
-  providers: ContractProviders<C, PCK>,
-  circuitId: PCK,
-  compiledContract: CompiledContract.CompiledContract<C, any>, // eslint-disable-line @typescript-eslint/no-explicit-any
-  contractAddress: ContractAddress
-): CircuitMaintenanceTxInterface => {
-  assertIsContractAddress(contractAddress);
-  return {
-    removeVerifierKey(): Promise<FinalizedTxData> {
-      return submitRemoveVerifierKeyTx(providers, compiledContract, contractAddress, circuitId);
-    },
-    insertVerifierKey(newVk: VerifierKey): Promise<FinalizedTxData> {
-      return submitInsertVerifierKeyTx(providers, compiledContract, contractAddress, circuitId, newVk);
-    }
-  };
-};
-
-/**
- * A set of maintenance transaction creation interfaces, one for each circuit defined in
- * a given contract, keyed by the circuit name.
- */
-export type CircuitMaintenanceTxInterfaces<C extends Contract.Any> = Record<Contract.ProvableCircuitId<C>, CircuitMaintenanceTxInterface>;
-
-/**
- * Creates a {@link CircuitMaintenanceTxInterfaces}.
- *
- * @param providers The providers to use to build transactions.
- * @param compiledContract The contract to use to execute circuits.
- * @param contractAddress The ledger address of the contract.
- */
-export const createCircuitMaintenanceTxInterfaces = <C extends Contract.Any>(
-  providers: ContractProviders<C>,
-  compiledContract: CompiledContract.CompiledContract<C, any>, // eslint-disable-line @typescript-eslint/no-explicit-any
-  contractAddress: ContractAddress
-): CircuitMaintenanceTxInterfaces<C> => {
-  assertIsContractAddress(contractAddress);
-  return ContractExecutable.make(compiledContract).getProvableCircuitIds().reduce(
-    (acc, circuitId) => ({
-      ...acc,
-      [circuitId]: createCircuitMaintenanceTxInterface(providers, circuitId, compiledContract, contractAddress)
-    }),
-    {}
-  ) as CircuitMaintenanceTxInterfaces<C>;
-};
-
-/**
- * Interface for creating maintenance transactions for a contract that was
- * deployed.
- */
-export interface ContractMaintenanceTxInterface {
-  /**
-   * Constructs and submits a transaction that replaces the maintenance
-   * authority stored on the blockchain for this contract.
-   *
-   * @param newAuthority The new contract maintenance authority for this contract.
-   */
-  replaceAuthority(newAuthority: SigningKey): Promise<FinalizedTxData>;
-}
-
-/**
- * Creates a {@link ContractMaintenanceTxInterface}.
- *
- * @param providers The providers to use to build transactions.
- * @param contractAddress The ledger address of the contract.
- */
-export const createContractMaintenanceTxInterface = <C extends Contract.Any>(
-  providers: ContractProviders,
-  compiledContract: CompiledContract.CompiledContract<C, any>, // eslint-disable-line @typescript-eslint/no-explicit-any
-  contractAddress: ContractAddress
-): ContractMaintenanceTxInterface => {
-  assertIsContractAddress(contractAddress);
-  return {
-    replaceAuthority: submitReplaceAuthorityTx(providers, compiledContract, contractAddress)
-  };
 };
